+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Add new `ProjectCreatorCard` component.
 - Feature: Add new `ClockIcon` component.
 - Feature: Add new `RightArrowIcon` component.
+- Feature: Add config & toolbar to default props `RichTextInput` component.
 - Fix: Fix display element as a block element (`ButtonImageWithText`, `ButtonImage`,
        `ButtonWithBadge`, `IconBadge`).
 - Fix: Remove `React.PropTypes` and use `PropTypes` from